+++ resources/views/event/overview.blade.php
@@ -52,5 +52,7 @@
                     @endphp
                     @include('layouts.alert')
                 @endforelse
+                {{ $events->links() }}
     </table>
+    {{ $events->links() }}
 @endsection