+++ resources/views/home.blade.php
@@ -1,16 +1,14 @@
 @extends('layouts.app')
 
 @section('content')
-    <h3>Welcome to the Rotterdam World Port – September 30th</h3>
-    @include('layouts.alert')
-    <p>On <strong>Sunday</strong>, the <strong>30th of September</strong>, the Dutch VACC will be hosting <strong>Rotterdam
-            World Port</strong>, a <strong>real operations & fly-in event</strong> at <strong>Rotterdam The Hague
-            Airport (EHRD)</strong>. Pilots are encouraged to book a real world operated flight through our booking
-        system. There is <strong>no specific arrival or departure time</strong> for booked flights - just feel free to
-        pop up or drop by whenever you like!</p>
-
-    <p>Rotterdam The Hague is the <strong>3rd largest airport</strong> in the Netherlands. It is extensively used by
-        Transavia which offers scheduled and charter flights to a great number of European destinations. The airport is
-        also regularly visited by VFR flights and bizzjets. We'll be open between <strong>17:30</strong> and <strong>20:00
-            UTC</strong>. Feel free to drop by on a commercial, bizzjet or VFR-flight!</p>
+    @if($event)
+        <h3>Welcome to {{ $event->name }} – {{ $event->startEvent->toFormattedDateString() }}</h3>
+        @include('layouts.alert')
+        @if($event->image_url)
+            <img src="{{ $event->image_url }}">
+        @endif
+        {!! $event->description !!}
+    @else
+        Currently no events scheduled.
+    @endif
 @endsection