+++ CHANGELOG.md
@@ -11,6 +11,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - ``Groupflight`` in ``EventType`` Enum and ``event_types`` table
+- The following event variables in ``events`` table
+    - ``import_only``
+    - ``uses_times``
+    - ``multiple_bookings_allowed``
+    - ``is_oceanic_event``
+- Event details on homepage if active
+- ``image_url`` in ``events`` table
+
+### Changed
+- Homepage now checks if event is active, and fill the title, date, and description
 
 ## [v0.6.3](https://gitlab.com/daveroverts/Book-me-a-cookie/compare/v0.6.2...v0.6.3) - 2018-09-29